change signature style for `useQuery` to conditional interface
